feat: cleanup app_context
